FRESH-334 : Modified DAO methods to also take ID into account
DAO methods were modified in order to take orgID into account when
retrieving bp products. In case none is found for the given org, there
is a fallback on the org 0.
Also deleted a methos that was not used anywhere and modified the
implementation and callings for another one which was wrong.