MAJOR hotfix on Trx commitNative/rollbackNative
Those 2 methods shall return true in case the transaction was not really
used (i.e. m_connection is null).

If we are not doing this, the afterCommit/afterRollback trx listeners
will not be called in case the transaction was not really used.

(cherry picked from commit d09352228759587f57d0c0f8dcc5a4848645ee57)