Porting different code pices to metasfresh
ArchiveEvent framework:
* add support to notify the system about a voided (also reversed etc)
document. That way, if it's not yet printed, it won't be printed.
  * this is implemented via an async processor
  * it's invoked via DocOutboundProducerValidator
* add suppord to specify the C_Doc_Outbound_Log_Line.Action on
firePdfUpdate()

DocActionBL: add general isReversalDocument() method

DocAction: add constants for common document column names

DefaultModelArchive: forward the async-batch-id from the archive
document to the archive-record

RecreateWorkPackageProcessor: add the code and a migration script to
insert it into the DB if it does not yet exist

Dunning: forward the async-batch-id from the dunning context to the
dunning document

1771: add legacy printing and archiving features

Task-Url: https://github.com/metasfresh/metasfresh/issues/1771