Augment IExpression to return CtxNames instead of strings as parameters
* IExpression: rename getParameters() to getParamertsAsPlainStrings and
added a new method getParameters() that returns CtxNames.
CtxNames are more useful because they can contain a default value, but
they can also contain a "normal" string.
In our case, we need CtxNames so we can evaluate the @CtxVar/Default@
notation that is contained in many of the expressions we deal with.
* CtxNames: add new static utility class around CtxNames to separate the
static code from the actual instance code. 
* Implement the new CtxName returning getParameters() method wherever
needed; in most cases that means returning the CtxNames that are already
embedded in the respective implementations.

Lookup validation rule shall support @CtxName/Default@ annotation
https://github.com/metasfresh/metasfresh-webui-api/issues/585