Implement attribute-based config for shipmentschedule handlers and..
use it for aggregation keys.

**DISCLAIMER**
We are not yet using it to create the shipment line's ASI! I'll create a follow-up for that.
Currently we only use it to decide if the attributes of two different HUs that were picked to the same shipment schedule mean that this schedule needs to end up ad two shipment lines.

some details:
* add attribute related config to shipment schedule handler (table `M_IolCandHandler`)
* extend ShipmentScheduleHandlerBL to get a schedule's handler
* make IShipmentScheduleHandler and abstract class (so i can add the attribute-config logic) and rename it to ShipmentScheduleHandler
* insert `M_ShipmentSchedule_Attributeconfig` record for all attributes that are in the "template" packing instruction
* make ShipmentScheduleWithHU use it (instead of `attributeValue.isDefinedByTemplate()` ) to compute a proper attribute-based aggregation key

make shipment HU aggregation configurable https://github.com/metasfresh/metasfresh/issues/3575