Return even inactive C_BPartner_Locations at many places (#10848)
Return even inactive C_BPartner_Locations at many places
Because e.g. if you have a completed invoice and someone deactivated the bpartner-location that the invoice refers to, then the invoice should still be able to e.g. get the location's country.
The same argument applies to a surprising number of documents.