Fix SQL problem about exporting INVOIC
The view on the customer-DB lacked the column "SiteName". But IDK why..the SQL 5624540_sys_gh12476_extend_and_change_invoice_edi.sql which was supposed to add it, was run.
Here, i'm creating the view with everything we have on master (additional minor things), to avoid problems when this is merged towards master.