Fix SQL problem about exporting INVOIC (#18119)
The views on the customer-DB lacked the column "SiteName". But IDK why..the SQL 5624540_sys_gh12476_extend_and_change_invoice_edi.sql and 5636240_sys_gh12886_edi_cctop_invoic_500_v.sql, which were supposed to add it, was run.

Here, i'm creating the views with everything we have on master (additional minor things), to avoid problems when this is merged towards master.